Switch where config files cause an issue
