+++ test/unit/module/runner/test_rule_configuration.py
@@ -7,7 +7,8 @@ import os
 from test.testlib.testcase import BaseTestCase
 
 from cfnlint.config import ConfigMixIn
-from cfnlint.runner import Runner, UnexpectedRuleException
+from cfnlint.exceptions import UnexpectedRuleException
+from cfnlint.runner import Runner
 
 
 class TestGetRules(BaseTestCase):
